refactor(markdown): extract reusable live editor core
Made-with: Proma
